merge sort of a vector
a full merge sort algorithm with vector splitting and recursion. Given a vector, it is split in two vectors and the sub-vectors are split until their length is less than 3, later we sort the sub-vectors by comparing two left values, swap them if needed, then apply merge sort for the other parts. This algorithm uses divide-and-conquer principle to sort small partitions of the vector and then merge them in sorted order.